Add onclick event for unit polygon
