+++ src/views/MapView/components/UnitGeometry/UnitGeometry.js
@@ -1,11 +1,49 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import PropTypes from 'prop-types';
+import swapCoordinates from '../../utils/swapCoordinates';
+import UnitHelper from '../../../../utils/unitHelper';
 
 
 const UnitGeometry = ({
-  geometryData,
+  data,
+  currentPage,
+  navigator,
 }) => {
   const { Polyline, Polygon } = global.rL;
+
+  const [geometryData, setGeometryData] = useState(null);
+
+  useEffect(() => {
+    const getUnitGeometry = (unit) => {
+      if ((currentPage === 'unit' || currentPage === 'search' || currentPage === 'fullList' || currentPage === 'event')) {
+        const { geometry } = unit;
+        if (geometry) {
+          const { coordinates } = geometry;
+          let unitGeometry;
+          switch (geometry.type) {
+            case 'MultiLineString': {
+              unitGeometry = swapCoordinates(coordinates);
+              break;
+            }
+            case 'MultiPolygon': {
+              unitGeometry = swapCoordinates(coordinates[0]);
+              break;
+            }
+            default:
+              return null;
+          }
+          return {
+            ...geometry,
+            coordinates: unitGeometry,
+          };
+        }
+      }
+      return null;
+    };
+
+    setGeometryData(getUnitGeometry(data));
+  }, [data]);
+
   const renderPolyline = () => {
     if (geometryData?.type !== 'MultiLineString' || !geometryData?.coordinates) {
       return null;
@@ -31,6 +69,15 @@ const UnitGeometry = ({
         positions={geometryData.coordinates}
         color="#ff8400"
         fillColor="#000"
+        onClick={(e) => {
+          try {
+            e.originalEvent.preventDefault();
+            e.originalEvent.stopPropagation();
+            UnitHelper.unitElementClick(navigator, data);
+          } catch (e) {
+            console.warn('Unable to process UnitGeometry Polygon click event');
+          }
+        }}
       />
     );
   };
@@ -48,10 +95,11 @@ const UnitGeometry = ({
 };
 
 UnitGeometry.propTypes = {
-  geometryData: PropTypes.shape({
-    coordinates: PropTypes.arrayOf(PropTypes.array).isRequired,
-    type: PropTypes.string.isRequired,
+  data: PropTypes.shape({
+    geometry: PropTypes.objectOf(PropTypes.any),
   }).isRequired,
+  currentPage: PropTypes.string.isRequired,
+  navigator: PropTypes.objectOf(PropTypes.any).isRequired,
 };
