+++ src/views/MapView/components/UnitGeometry/index.js
@@ -1,3 +1,17 @@
+import { connect } from 'react-redux';
 import UnitGeometry from './UnitGeometry';
 
-export default UnitGeometry;
+const mapStateToProps = (state) => {
+  const {
+    navigator, user,
+  } = state;
+  const {
+    page,
+  } = user;
+  return {
+    currentPage: page,
+    navigator,
+  };
+};
+
+export default connect(mapStateToProps)(UnitGeometry);
